feat: OData Query (initial tests)
