+++ src/CommunityToolkit.Datasync.Client/CommunityToolkit.Datasync.Client.csproj
@@ -10,5 +10,6 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.7" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Azure.Core.Spatial" Version="1.1.0" />
   </ItemGroup>
 </Project>
