+++ src/CommunityToolkit.Datasync.Client/Query/Linq/FilterBuildingExpressionVisitor.cs
@@ -2,12 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-// Several methods can be static, but are also generic.  We have to choose which
-// warning to disable, since there will always be a warning.  I've opted to make
-// the methods a part of the class rather than static throughout the class.
-#pragma warning disable CA1822 // Mark members as static
+// Reflection and LINQ requires a lot of null manipulation, so we opt for
+// a generalized "nullable" option here to allow us to do that.
+#nullable disable
 
-// Long switch statements are easier to read than switch expressions with pattern matching.
+// Stylistically, this file has a couple of switch statements that are long,
+// and we've opted for the more readable version of these switch statements.
 #pragma warning disable IDE0066 // Convert switch statement to expression
 
 using System.Diagnostics.CodeAnalysis;
@@ -15,31 +15,15 @@ using System.Linq.Expressions;
 using System.Reflection;
 using System.Text.Json;
 
-namespace CommunityToolkit.Datasync.Client.Query;
+namespace CommunityToolkit.Datasync.Client.Query.Linq;
 
 /// <summary>
-/// Walks the LINQ expression tree to compile the expression into a simplified tree of nodes that can be used for OData or SQL production.
+/// Walks the LINQ expression tree to compile the expression into a simplified tree of
+/// nodes that can be used for OData query string production.
 /// </summary>
-internal class FilterBuildingExpressionVisitor
+internal sealed class FilterBuildingExpressionVisitor
 {
-    // string.Concat(string, string)
-    private static readonly MethodInfo concatMethod = typeof(string).GetRuntimeMethod("Concat", [typeof(string), typeof(string)])!;
-    // string.StartsWith(string, StringComparison)
-    private static readonly MethodInfo startsWithMethod = typeof(string).GetRuntimeMethod("StartsWith", [typeof(string), typeof(StringComparison)])!;
-    // string.EndsWith(string, StringComparison)
-    private static readonly MethodInfo endsWithMethod = typeof(string).GetRuntimeMethod("EndsWith", [typeof(string), typeof(StringComparison)])!;
-    // string.Equals(string)
-    private static readonly MethodInfo equals1Method = typeof(string).GetRuntimeMethod("Equals", [typeof(string)])!;
-    // string.Equals(string, StringComparison)
-    private static readonly MethodInfo equals2Method = typeof(string).GetRuntimeMethod("Equals", [typeof(string), typeof(StringComparison)])!;
-    // string.ToString()
-    private static readonly MethodInfo toStringMethod = typeof(object).GetTypeInfo().GetDeclaredMethod("ToString")!;
-
-    // IEnumerable<string>.Contains(string) - generic extension method
-    private static readonly MethodInfo arrayContainsMethod = typeof(Enumerable).GetRuntimeMethods()
-        .Single(m => m.Name == "Contains" && m.GetParameters().Length == 2)
-        .MakeGenericMethod(typeof(string));
-
+    #region DO NOT DELETE THIS SECTION
     /// <summary>
     /// <para>** Do not remove this code. **</para>
     /// </summary>
@@ -51,76 +35,86 @@ internal class FilterBuildingExpressionVisitor
     /// reflection, so we need this code here to ensure that don't get removed by the compiler.
     /// </remarks>
     [SuppressMessage("Style", "IDE0059:Unnecessary assignment of a value", Justification = "See code comments - this is required!!!")]
-    [SuppressMessage("Style", "IDE0049:Simplify Names", Justification = "<Pending>")]
-#pragma warning disable CS8600 // Converting null literal or possible null value to non-nullable type.
     static FilterBuildingExpressionVisitor()
     {
-        string aString = new Object().ToString();
+        string aString = new object().ToString();
         aString = string.Concat(aString, "a string");
     }
-#pragma warning restore CS8600 // Converting null literal or possible null value to non-nullable type.
+    #endregion
+
+    // MethodInfo for string.Concat(string, string)
+    private static readonly MethodInfo concatMethod = typeof(string).GetRuntimeMethod("Concat", [typeof(string), typeof(string)]);
+    // MethodInfo for string.StartsWith(string, StringComparison)
+    private static readonly MethodInfo startsWithMethod = typeof(string).GetRuntimeMethod("StartsWith", [typeof(string), typeof(StringComparison)]);
+    // MethodInfo for string.EndsWith(string, StringComparison)
+    private static readonly MethodInfo endsWithMethod = typeof(string).GetRuntimeMethod("EndsWith", [typeof(string), typeof(StringComparison)]);
+    // MethodInfo for string.Equals(string)
+    private static readonly MethodInfo equals1Method = typeof(string).GetRuntimeMethod("Equals", [typeof(string)]);
+    // MethodInfo for string.Equals(string, StringComparison)
+    private static readonly MethodInfo equals2Method = typeof(string).GetRuntimeMethod("Equals", [typeof(string), typeof(StringComparison)]);
+    // MethodInfo for IEnumerable<string>.Contains(string) - generic extension method
+    private static readonly MethodInfo arrayContainsMethod = typeof(Enumerable).GetRuntimeMethods().Single(m => m.Name == "Contains" && m.GetParameters().Length == 2).MakeGenericMethod(typeof(string));
+    // MethodInfo for string.ToString()
+    private static readonly MethodInfo toStringMethod = typeof(object).GetTypeInfo().GetDeclaredMethod("ToString");
 
     /// <summary>
     /// Initializes a new instance of the <see cref="FilterBuildingExpressionVisitor"/>
     /// </summary>
-    /// <param name="serializerOptions">The <see cref="JsonSerializerOptions"/> to use for resolving names.</param>
-    internal FilterBuildingExpressionVisitor(JsonSerializerOptions serializerOptions) : base()
+    /// <param name="namingPolicy">The <see cref="JsonNamingPolicy"/> to use to determine property names from members used with expressions.</param>
+    internal FilterBuildingExpressionVisitor(JsonNamingPolicy namingPolicy) : base()
     {
-        SerializerOptions = serializerOptions;
+        ArgumentNullException.ThrowIfNull(namingPolicy, nameof(namingPolicy));
+        NamingPolicy = namingPolicy;
     }
 
     /// <summary>
-    /// The contract resolvers to use to determine property names from members used with expressions.
+    /// The naming policy to use in converting member names into a serialized property name.
     /// </summary>
-    internal JsonSerializerOptions SerializerOptions { get; }
+    internal JsonNamingPolicy NamingPolicy { get; }
 
     /// <summary>
-    /// The accumulator for the <see cref="QueryNode"/> being generated by the compiler
+    /// The accumulator for the <see cref="QueryNode"/> representing the filter in the expression tree.
     /// </summary>
     internal Stack<QueryNode> FilterExpression { get; } = new();
 
     /// <summary>
-    /// Translate an expression tree into a compiled OData query.
+    /// Compiles an expression tree representing the predicate for a filter into a query node stack.
     /// </summary>
-    /// <param name="expression">The expression tree to translate</param>
-    /// <param name="serializerOptions">The <see cref="JsonSerializerOptions"/> to use for resolving names.</param>
-    /// <returns>An OData <see cref="QueryNode"/></returns>
-    internal static QueryNode? Compile(Expression expression, JsonSerializerOptions serializerOptions)
+    /// <param name="filterExpression">The filter expression to compile.</param>
+    /// <param name="namingPolicy">The naming policy for converting member names into serialized property names.</param>
+    /// <returns>The root <see cref="QueryNode"/> for the compiled filter expression.</returns>
+    internal static QueryNode Compile(Expression filterExpression, JsonNamingPolicy namingPolicy)
     {
-        FilterBuildingExpressionVisitor visitor = new(serializerOptions);
-        _ = visitor.Visit(expression);
+        FilterBuildingExpressionVisitor visitor = new(namingPolicy);
+        _ = visitor.Visit(filterExpression);
         return visitor.FilterExpression.FirstOrDefault();
-    }
-
-    /// <summary>
-    /// Converts the name provided according to the property naming policy.  If a property naming policy
-    /// is not set, then the name is returned.
-    /// </summary>
-    /// <param name="name">The name to be converted.</param>
-    /// <param name="serializerOptions">The serializer options to use for converting the name.</param>
-    /// <returns>The converted name.</returns>
-    internal static string ConvertName(string name, JsonSerializerOptions serializerOptions)
-    {
-        return serializerOptions.PropertyNamingPolicy?.ConvertName(name) ?? name;
     }
 
     /// <summary>
     /// Gets the table member name referenced by an expression, or return null.
     /// </summary>
     /// <param name="expression">The expression to check</param>
-    /// <param name="serializerOptions">The <see cref="JsonSerializerOptions"/> to use for resolving names.</param>
-    /// <returns>The table member name</returns>
-    internal static string GetTableMemberName(Expression expression, JsonSerializerOptions serializerOptions)
+    /// <param name="namingPolicy">The naming policy for converting member names into serialized property names.</param>
+    /// <returns>the table member name</returns>
+    internal static string GetTableMemberName(Expression expression, JsonNamingPolicy namingPolicy)
     {
-        if (expression is MemberExpression member && member.Expression != null)
+        ArgumentNullException.ThrowIfNull(expression, nameof(expression));
+        ArgumentNullException.ThrowIfNull(namingPolicy, nameof(namingPolicy));
+
+        if (expression is MemberExpression member)
         {
-            if (member.Expression.NodeType == ExpressionType.Parameter || (member.Expression.NodeType == ExpressionType.Convert && member.Member.MemberType == MemberTypes.Property))
+            if (member.Expression.NodeType == ExpressionType.Parameter)
             {
-                return ConvertName(member.Member.Name, serializerOptions);
+                return namingPolicy?.ConvertName(member.Member.Name) ?? member.Member.Name;
+            }
+
+            if (member.Expression.NodeType == ExpressionType.Convert && member.Member.MemberType == MemberTypes.Property)
+            {
+                return namingPolicy?.ConvertName(member.Member.Name) ?? member.Member.Name;
             }
         }
 
-        return string.Empty;
+        return null;
     }
 
     /// <summary>
@@ -130,6 +124,11 @@ internal class FilterBuildingExpressionVisitor
     /// <returns>The visited node.</returns>
     internal Expression Visit(Expression node)
     {
+        if (node == null)
+        {
+            return null;
+        }
+
         switch (node.NodeType)
         {
             case ExpressionType.Add:
@@ -164,7 +163,7 @@ internal class FilterBuildingExpressionVisitor
             case ExpressionType.Call:
                 return VisitMethodCallExpression((MethodCallExpression)node);
 
-            case ExpressionType.Convert:
+            case ExpressionType.Convert: // TODO: Validate that this is the correct mechanism and not a ConvertExpression instead!
             case ExpressionType.Not:
             case ExpressionType.Quote:
                 return VisitUnaryExpression((UnaryExpression)node);
@@ -181,10 +180,10 @@ internal class FilterBuildingExpressionVisitor
     /// <returns>The visited expression</returns>
     internal Expression VisitBinaryExpression(BinaryExpression node)
     {
-        if (IsEnumExpression(node, out UnaryExpression? enumExpression, out ConstantExpression? constant))
+        if (IsEnumExpression(node, out UnaryExpression enumExpression, out ConstantExpression constant) && enumExpression != null && constant != null)
         {
             Type enumType = enumExpression.Operand.Type;
-            string enumString = Enum.ToObject(enumType, constant.Value!).ToString()!;
+            string enumString = Enum.ToObject(enumType, constant.Value).ToString();
             Expression call = Expression.Call(enumExpression, toStringMethod);
             _ = Visit(Expression.MakeBinary(node.NodeType, call, Expression.Constant(enumString)));
         }
@@ -211,7 +210,7 @@ internal class FilterBuildingExpressionVisitor
     /// <returns>The visited expression</returns>
     private ConstantExpression VisitConstantExpression(ConstantExpression node)
     {
-        FilterExpression.Push(new ConstantNode(node.Value!));
+        FilterExpression.Push(new ConstantNode(node.Value));
         return node;
     }
 
@@ -223,21 +222,20 @@ internal class FilterBuildingExpressionVisitor
     private MemberExpression VisitMemberExpression(MemberExpression node)
     {
         // Is the member the name of a member?
-        string memberName = GetTableMemberName(node, SerializerOptions);
-        if (!string.IsNullOrEmpty(memberName))
+        string memberName = GetTableMemberName(node, NamingPolicy);
+        if (memberName != null)
         {
             FilterExpression.Push(new MemberAccessNode(null, memberName));
             return node;
         }
 
         // Is this member actually a function that looks like a property (e.g. string.Length)
-        MemberInfoKey key = new(node.Member);
-        string? methodName = InstanceProperties.GetMethodName(key);
+        string methodName = InstanceProperties.GetMethodName(new MemberInfoKey(node.Member));
         if (methodName != null)
         {
             FunctionCallNode fnCallNode = new(methodName);
             FilterExpression.Push(fnCallNode);
-            _ = Visit(node.Expression!);
+            _ = Visit(node.Expression);
             SetChildren(fnCallNode);
             return node;
         }
@@ -253,94 +251,90 @@ internal class FilterBuildingExpressionVisitor
     /// <returns>The visited node</returns>
     internal Expression VisitMethodCallExpression(MethodCallExpression node)
     {
-        MethodInfo? baseDefinition = node.Method.GetRuntimeBaseDefinition();
+        MethodInfo baseDefinition = node.Method.GetRuntimeBaseDefinition();
 
-        if (baseDefinition != null)
+        //
+        // string.StartsWith(string, StringComparison)
+        //
+        if (baseDefinition.Equals(startsWithMethod) && node.Arguments.Count == 2 && node.Arguments[1].Type == typeof(StringComparison))
         {
-            //
-            // string.StartsWith(string, StringComparison)
-            //
-            if (baseDefinition.Equals(startsWithMethod) && node.Arguments.Count == 2 && node.Arguments[1].Type == typeof(StringComparison))
+            FunctionCallNode callNode = new("startswith");
+            FilterExpression.Push(callNode);
+            _ = VisitStringComparisonExpression(callNode, node);
+            return node;
+        }
+
+        //
+        // string.EndsWith(string, StringComparison)
+        //
+        if (baseDefinition.Equals(endsWithMethod) && node.Arguments.Count == 2 && node.Arguments[1].Type == typeof(StringComparison))
+        {
+            FunctionCallNode callNode = new("endswith");
+            FilterExpression.Push(callNode);
+            _ = VisitStringComparisonExpression(callNode, node);
+            return node;
+        }
+
+        //
+        // string.Equals(string)
+        //
+        if (baseDefinition.Equals(equals1Method) && node.Arguments.Count == 1)
+        {
+            BinaryOperatorNode equalityNode = new(BinaryOperatorKind.Equal);
+            FilterExpression.Push(equalityNode);
+            _ = Visit(node.Object);
+            _ = Visit(node.Arguments[0]);
+            SetChildren(equalityNode);
+            return node;
+        }
+
+        //
+        // string.Equals(string, StringComparison)
+        //
+        if (baseDefinition.Equals(equals2Method) && node.Arguments.Count == 2 && node.Arguments[1].Type == typeof(StringComparison))
+        {
+            BinaryOperatorNode equalityNode = new(BinaryOperatorKind.Equal);
+            FilterExpression.Push(equalityNode);
+            _ = VisitStringComparisonExpression(equalityNode, node);
+            return node;
+        }
+
+        //
+        // string[].Contains(string)
+        //
+        if (baseDefinition.Equals(arrayContainsMethod) && node.Arguments.Count == 2 && node.Arguments[0].Type == typeof(string[]))
+        {
+            FunctionCallNode callNode = new("in");
+            FilterExpression.Push(callNode);
+            _ = Visit(node.Arguments[0]);
+            _ = Visit(node.Arguments[1]);
+            SetChildren(callNode);
+            return node;
+        }
+
+        //
+        // Anything in the MethodNames table (which has a direct relationship to the OData representation)
+        //
+        if (MethodNames.TryGetValue(new MemberInfoKey(node.Method), out string methodName, out bool isStatic))
+        {
+            FunctionCallNode fnCallNode = new(methodName);
+            FilterExpression.Push(fnCallNode);
+            foreach (Expression argument in (isStatic ? [] : new Expression[] { node.Object }).Concat(node.Arguments))
             {
-                FunctionCallNode callNode = new("startswith");
-                FilterExpression.Push(callNode);
-                _ = VisitStringComparisonExpression(callNode, node);
-                return node;
+                _ = Visit(argument);
             }
 
-            //
-            // string.EndsWith(string, StringComparison)
-            //
-            if (baseDefinition.Equals(endsWithMethod) && node.Arguments.Count == 2 && node.Arguments[1].Type == typeof(StringComparison))
-            {
-                FunctionCallNode callNode = new("endswith");
-                FilterExpression.Push(callNode);
-                _ = VisitStringComparisonExpression(callNode, node);
-                return node;
-            }
+            SetChildren(fnCallNode);
+            return node;
+        }
 
-            //
-            // string.Equals(string)
-            //
-            if (baseDefinition.Equals(equals1Method) && node.Arguments.Count == 1)
-            {
-                BinaryOperatorNode equalityNode = new(BinaryOperatorKind.Equal);
-                FilterExpression.Push(equalityNode);
-                _ = Visit(node.Object!);
-                _ = Visit(node.Arguments[0]);
-                SetChildren(equalityNode);
-                return node;
-            }
-
-            //
-            // string.Equals(string, StringComparison)
-            //
-            if (baseDefinition.Equals(equals2Method) && node.Arguments.Count == 2 && node.Arguments[1].Type == typeof(StringComparison))
-            {
-                BinaryOperatorNode equalityNode = new(BinaryOperatorKind.Equal);
-                FilterExpression.Push(equalityNode);
-                _ = VisitStringComparisonExpression(equalityNode, node);
-                return node;
-            }
-
-            //
-            // string[].Contains(string)
-            //
-            if (baseDefinition.Equals(arrayContainsMethod) && node.Arguments.Count == 2 && node.Arguments[0].Type == typeof(string[]))
-            {
-                FunctionCallNode callNode = new("in");
-                FilterExpression.Push(callNode);
-                _ = Visit(node.Arguments[0]);
-                _ = Visit(node.Arguments[1]);
-                SetChildren(callNode);
-                return node;
-            }
-
-            //
-            // Anything in the MethodNames table (which has a direct relationship to the OData representation)
-            //
-            MemberInfoKey key = new(node.Method);
-            if (MethodNames.TryGetValue(key, out string? methodName, out bool isStatic))
-            {
-                FunctionCallNode fnCallNode = new(methodName);
-                FilterExpression.Push(fnCallNode);
-                foreach (Expression? argument in (isStatic ? [] : new Expression[] { node.Object! }).Concat(node.Arguments))
-                {
-                    _ = Visit(argument);
-                }
-
-                SetChildren(fnCallNode);
-                return node;
-            }
-
-            //
-            // string.ToString()
-            //
-            if (baseDefinition.Equals(toStringMethod))
-            {
-                _ = Visit(node.Object!);
-                return node;
-            }
+        //
+        // string.ToString()
+        //
+        if (baseDefinition.Equals(toStringMethod))
+        {
+            _ = Visit(node.Object);
+            return node;
         }
 
         //
@@ -352,8 +346,8 @@ internal class FilterBuildingExpressionVisitor
     /// <summary>
     /// Process a two-argument string-expression (startsWith, endsWith)
     /// </summary>
-    /// <param name="queryNode">The query node that we are creating.</param>
-    /// <param name="node">The expression to visit</param>
+    /// <param name="queryNode">The <see cref="QueryNode"/> being created.</param>
+    /// <param name="node">The node to visit</param>
     /// <returns>The visited node</returns>
     internal Expression VisitStringComparisonExpression(QueryNode queryNode, MethodCallExpression node)
     {
@@ -362,7 +356,7 @@ internal class FilterBuildingExpressionVisitor
         {
             case StringComparison.Ordinal:
             case StringComparison.InvariantCulture:
-                _ = Visit(node.Object!);
+                _ = Visit(node.Object);
                 _ = Visit(node.Arguments[0]);
                 SetChildren(queryNode);
                 return node;
@@ -371,7 +365,7 @@ internal class FilterBuildingExpressionVisitor
             case StringComparison.InvariantCultureIgnoreCase:
                 FunctionCallNode arg1 = new("tolower");
                 FilterExpression.Push(arg1);
-                _ = Visit(node.Object!);
+                _ = Visit(node.Object);
                 SetChildren(arg1);
 
                 FunctionCallNode arg2 = new("tolower");
@@ -395,8 +389,7 @@ internal class FilterBuildingExpressionVisitor
     /// <returns>The StringComparison value.</returns>
     internal StringComparison GetStringComparisonFromExpression(Expression node)
     {
-        // Get the StringComparison argument by visiting the node.
-        // Don't leave it on the stack!
+        // Get the StringComparison argument by visiting the node, but don't leave it on the stack
         _ = Visit(node);
         QueryNode comparisonNode = FilterExpression.Pop();
 
@@ -449,7 +442,7 @@ internal class FilterBuildingExpressionVisitor
     /// <param name="to">The type to convert to</param>
     /// <returns>True if there is an implicit conversion</returns>
     internal bool IsConversionImplicit(UnaryExpression node, Type from, Type to)
-        => GetTableMemberName(node.Operand, SerializerOptions) != null && ImplicitConversions.IsImplicitConversion(from, to);
+        => GetTableMemberName(node.Operand, NamingPolicy) != null && ImplicitConversions.IsImplicitConversion(from, to);
 
     /// <summary>
     /// Checks if the provided binary expression is an enum.
@@ -458,7 +451,7 @@ internal class FilterBuildingExpressionVisitor
     /// <param name="unaryExpression">The expression which is the enum.</param>
     /// <param name="constantExpression">The constant expression containing the enum value</param>
     /// <returns>Trye if an enum expression is found.</returns>
-    internal bool IsEnumExpression(BinaryExpression node, [NotNullWhen(true)] out UnaryExpression? unaryExpression, [NotNullWhen(true)] out ConstantExpression? constantExpression)
+    internal static bool IsEnumExpression(BinaryExpression node, out UnaryExpression unaryExpression, out ConstantExpression constantExpression)
     {
         // Case 1: enum on the left side
         if (node.Left is UnaryExpression left && IsEnumExpression(left) && node.Right is ConstantExpression leftExpr)
@@ -502,6 +495,6 @@ internal class FilterBuildingExpressionVisitor
             args.Push(FilterExpression.Pop());
         }
 
-        parent.SetChildren([..args]);
+        parent.SetChildren([.. args]);
     }
 }
