+++ src/CommunityToolkit.Datasync.Client/Query/Linq/QueryNodeVisitor.cs
@@ -2,7 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace CommunityToolkit.Datasync.Client.Query;
+// Reflection and LINQ requires a lot of null manipulation, so we opt for
+// a generalized "nullable" option here to allow us to do that.
+#nullable disable
+
+namespace CommunityToolkit.Datasync.Client.Query.Linq;
 
 /// <summary>
 /// Visitor interface for walking the <see cref="QueryNode"/> tree.
@@ -13,30 +17,30 @@ internal abstract class QueryNodeVisitor<T>
     /// <summary>
     /// Visit a <see cref="BinaryOperatorNode"/>
     /// </summary>
-    internal abstract T Visit(BinaryOperatorNode node);
+    public abstract T Visit(BinaryOperatorNode node);
 
     /// <summary>
     /// Visit a <see cref="ConstantNode"/>
     /// </summary>
-    internal abstract T Visit(ConstantNode node);
+    public abstract T Visit(ConstantNode node);
 
     /// <summary>
     /// Visit a <see cref="ConvertNode"/>
     /// </summary>
-    internal abstract T Visit(ConvertNode node);
+    public abstract T Visit(ConvertNode node);
 
     /// <summary>
     /// Visit a <see cref="FunctionCallNode"/>
     /// </summary>
-    internal abstract T Visit(FunctionCallNode node);
+    public abstract T Visit(FunctionCallNode node);
 
     /// <summary>
     /// Visit a <see cref="MemberAccessNode"/>
     /// </summary>
-    internal abstract T Visit(MemberAccessNode node);
+    public abstract T Visit(MemberAccessNode node);
 
     /// <summary>
     /// Visit a <see cref="UnaryOperatorNode"/>
     /// </summary>
-    internal abstract T Visit(UnaryOperatorNode node);
+    public abstract T Visit(UnaryOperatorNode node);
 }
