--- src/CommunityToolkit.Datasync.Client/Query/UnaryOperatorKind.cs
@@ -1,14 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace CommunityToolkit.Datasync.Client.Query;
-
-/// <summary>
-/// Enumeration of the supported unary operators.
-/// </summary>
-internal enum UnaryOperatorKind
-{
-    Negate,
-    Not
-}
